requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
aiosqlite
alembic
annotated-types
anyio
apache-airflow
apache-airflow-providers-common-compat
apache-airflow-providers-common-io
apache-airflow-providers-common-sql
apache-airflow-providers-fab
apache-airflow-providers-ftp
apache-airflow-providers-http
apache-airflow-providers-imap
apache-airflow-providers-smtp
apache-airflow-providers-sqlite
apispec
argcomplete
argon2-cffi
argon2-cffi-bindings
asgiref
attrs
babel
beautifulsoup4
blinker
cachelib
certifi
cffi
charset-normalizer
click
clickclick
colorama
colorlog
ConfigUpdater
connexion
cramjam
cron-descriptor
croniter
cryptography
Deprecated
dill
dnspython
docling
docling-core
docling-ibm-models
docling-parse
easyocr
email_validator
et_xmlfile
fastparquet
filelock
filetype
Flask
Flask-AppBuilder
Flask-Babel
Flask-Caching
Flask-JWT-Extended
Flask-Limiter
Flask-Login
Flask-Session
Flask-SQLAlchemy
Flask-WTF
frozenlist
fsspec
google-re2
googleapis-common-protos
greenlet
grpcio
gunicorn
h11
httpcore
httpx
huggingface-hub
idna
imageio
importlib_metadata
inflection
itsdangerous
Jinja2
jmespath
jsonlines
jsonref
jsonschema
jsonschema-specifications
latex2mathml
lazy-object-proxy
lazy_loader
limits
linkify-it-py
lockfile
lxml
Mako
markdown-it-py
marko
MarkupSafe
marshmallow
marshmallow-oneofschema
marshmallow-sqlalchemy
mdit-py-plugins
mdurl
methodtools
minio
more-itertools
mpire
mpmath
multidict
multiprocess
networkx
ninja
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
opencv-python-headless
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
ordered-set
packaging
pandas
pathspec
pendulum
pillow
pluggy
prison
propcache
protobuf
psutil
pyclipper
pycparser
pycryptodome
pydantic
pydantic-settings
pydantic_core
Pygments
PyJWT
pylatexenc
pypdfium2
python-bidi
python-daemon
python-dateutil
python-docx
python-dotenv
python-nvd3
python-pptx
python-slugify
pytz
PyYAML
referencing
regex
requests
requests-toolbelt
rfc3339-validator
rich
rich-argparse
rpds-py
rtree
safetensors
scikit-image
scipy
semchunk
setproctitle
setuptools
shapely
shellingham
six
sniffio
soupsieve
SQLAlchemy
SQLAlchemy-JSONField
SQLAlchemy-Utils
sqlparse
sympy
tabulate
tenacity
termcolor
text-unidecode
tifffile
time-machine
tokenizers
torch
torchvision
tqdm
transformers
triton
typer
typing-inspection
typing_extensions
tzdata
uc-micro-py
universal_pathlib
urllib3
Werkzeug
wirerope
wrapt
WTForms
XlsxWriter
yarl
zipp
